Created a scripts-modernizr task to rebuild the Modernizr feature detection script via Modernizr 3.
This script executes the `modernizr` command line installed via: `npm install -g modernizr`, rather than the
gulp-modernizr plugin so that we can control the tests we need via a JSON config file.

You probably shouldn't run this gulp task directly. Rather, run `gulp scripts-vendor`, and it will rebuild the
modernizr.js file and concatenate it with the rest of the vendor js scripts (those that need to be placed in the
<head> of the web page).

TODO: Create a test that warns you, nicely, that you need to globally install `modernizr` on your system first.